requirements.txt
# ============================================================
# 网络韧性优化框架 - 依赖项
# Neural-Symbolic Network Resilience Optimization Framework
# ============================================================

# ==================== 核心依赖 ====================
# PyTorch - 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# ==================== 图计算 ====================
# NetworkX - 图操作和分析
networkx>=3.0

# PyTorch Geometric - 图神经网络 (可选)
# 注意：需要根据 PyTorch 版本安装对应版本
# pip install torch_geometric
# pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv -f https://data.pyg.org/whl/torch-2.0.0+cu118.html
torch-geometric>=2.3.0

# ==================== 大语言模型 ====================
# HuggingFace Transformers
transformers>=4.35.0

# PEFT - 参数高效微调 (LoRA)
peft>=0.6.0

# Tokenizers
tokenizers>=0.14.0

# Accelerate - 分布式训练
accelerate>=0.24.0

# BitsAndBytes - 量化训练
bitsandbytes>=0.41.0

# ==================== 数据处理 ====================
# NumPy
numpy>=1.24.0

# Pandas
pandas>=2.0.0

# SciPy - 科学计算 (稀疏矩阵、特征值)
scipy>=1.10.0

# ==================== 配置和工具 ====================
# YAML 配置
pyyaml>=6.0

# 进度条
tqdm>=4.65.0

# 日志
loguru>=0.7.0

# ==================== 可视化 (可选) ====================
matplotlib>=3.7.0
seaborn>=0.12.0

# ==================== 开发工具 (可选) ====================
# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# 代码格式化
black>=23.0.0
isort>=5.12.0

# 类型检查
mypy>=1.0.0

# 测试
pytest>=7.0.0
pytest-cov>=4.0.0

# ==================== 其他工具 ====================
# 命令行参数解析
click>=8.0.0

# 环境变量管理
python-dotenv>=1.0.0
